requirements.txt
discord.py==1.3.4
discord-webhook==0.8.0